created Search api route

+++ server/controllers/listing.controller.js
@@ -1,11 +1,10 @@
 import Listing from "../models/listing.model.js";
+import { errorHandler } from "../utils/error.js";
 
 export const createListing = async (req, res, next) => {
-  const newListing = req.body;
-  const userId = req.user.id;
   try {
-    const listing = await Listing.create({ ...newListing, userRef: userId });
-    res.status(201).json(listing);
+    const listing = await Listing.create(req.body);
+    return res.status(201).json(listing);
   } catch (error) {
     next(error);
   }
@@ -18,9 +17,13 @@ export const deleteListing = async (req, res, next) => {
     return next(errorHandler(404, "Listing not found!"));
   }
 
+  if (req.user.id !== listing.userRef) {
+    return next(errorHandler(401, "You can only delete your own listings!"));
+  }
+
   try {
     await Listing.findByIdAndDelete(req.params.id);
-    res.status(200).json("Listing has been deleted");
+    res.status(200).json("Listing has been deleted!");
   } catch (error) {
     next(error);
   }
@@ -28,9 +31,11 @@ export const deleteListing = async (req, res, next) => {
 
 export const updateListing = async (req, res, next) => {
   const listing = await Listing.findById(req.params.id);
-
   if (!listing) {
     return next(errorHandler(404, "Listing not found!"));
+  }
+  if (req.user.id !== listing.userRef) {
+    return next(errorHandler(401, "You can only update your own listings!"));
   }
 
   try {
@@ -45,24 +50,75 @@ export const updateListing = async (req, res, next) => {
   }
 };
 
-export const getAllListings = async (req, res, next) => {
-  const userId = req.user.id;
-
+export const getListing = async (req, res, next) => {
   try {
-    const listings = await Listing.find({ userRef: userId });
+    const listing = await Listing.findById(req.params.id);
+    if (!listing) {
+      return next(errorHandler(404, "Listing not found!"));
+    }
+    res.status(200).json(listing);
+  } catch (error) {
+    next(error);
+  }
+};
+
+export const getAllListings = async (req, res, next) => {
+  try {
+    const listings = await Listing.find();
     res.status(200).json(listings);
   } catch (error) {
     next(error);
   }
 };
 
-export const getListing = async (req, res, next) => {
+export const getListings = async (req, res, next) => {
   try {
-    const listing = await Listing.findById(req.params.id);
+    const limit = parseInt(req.query.limit) || 9;
+    const startIndex = parseInt(req.query.startIndex) || 0;
+    let offer = req.query.offer;
 
-    if (!listing) return next(errorHandler(404, "Listing not found!"));
+    if (offer === undefined || offer === "false") {
+      offer = { $in: [false, true] };
+    }
 
-    res.status(200).json(listing);
+    let furnished = req.query.furnished;
+
+    if (furnished === undefined || furnished === "false") {
+      furnished = { $in: [false, true] };
+    }
+
+    let parking = req.query.parking;
+
+    if (parking === undefined || parking === "false") {
+      parking = { $in: [false, true] };
+    }
+
+    let type = req.query.type;
+
+    if (type === undefined || type === "all") {
+      type = { $in: ["sale", "rent"] };
+    }
+
+    const searchTerm = req.query.searchTerm || "";
+
+    const sort = req.query.sort || "createdAt";
+
+    const order = req.query.order || "desc";
+
+    const listings = await Listing.find({
+      name: { $regex: searchTerm, $options: "i" },
+      offer,
+      furnished,
+      parking,
+      type,
+    })
+      .sort({ [sort]: order })
+      .limit(limit)
+      .skip(startIndex);
+
+    console.log("listings : ", listings);
+
+    return res.status(200).json(listings);
   } catch (error) {
     next(error);
   }
